Updated structure web app and more
