+++ requirements-dev.txt
@@ -4,3 +4,5 @@ sphinx>=7.2.6
 pre-commit>=3.6.2
 sphinx-autobuild>=2024.2.4
 pytest>=8.1.1
+streamlit>=1.32.2
+python-dotenv>=1.0.1
