+++ requirements.txt
@@ -1,3 +1,5 @@
 setuptools>=65.5.0
 openai>=1.12.0
 pytest>=8.1.1
+streamlit>=1.32.2
+python-dotenv>=1.0.1
